Updated data model
The underlying data model has been rewritten so it can support things
like:
Divisions with an arbitrary number of teams
Divisions with names
Different default divisions based on the division count

Still to come: the division list does not update to the most recent
interaction, but the one previous to that.

Also: fixed a bug with the team layout that was causing conferences to
display vertically instead of horizontally.